refactor: scheduler end and begin of piece
