+++ pkg/rpc/scheduler/client/client.go
@@ -158,7 +158,7 @@ func (sc *schedulerClient) ReportPieceResult(ctx context.Context, taskID string,
 	logger.With("peerId", ptr.PeerId, "errMsg", err).Infof("start to report piece result for taskID: %s", taskID)
 
 	// trigger scheduling
-	return pps, pps.Send(scheduler.NewZeroPieceResult(taskID, ptr.PeerId))
+	return pps, pps.Send(NewBeginOfPiece(taskID, ptr.PeerId))
 }
 
 func (sc *schedulerClient) ReportPeerResult(ctx context.Context, pr *scheduler.PeerResult, opts ...grpc.CallOption) error {
